Pull Runner apart for testability and usability
* RepeatedMeasurement to unit test it and not pollute general
runner with this very specific code for runtime measurements
that's only relevant on Windows these days it seems
* Hooks code in its own module to not pollute and reuse from both
place
* made `measure` public on the runner... not 100% happy with this
but putting it into even another module seems like an overhead
now.
* temporarily duplicated `main_function` code
