+++ lib/benchee/benchmark/runner.ex
@@ -5,8 +5,8 @@ defmodule Benchee.Benchmark.Runner do
   """
 
   alias Benchee.Benchmark
-  alias Benchee.Benchmark.{Scenario, ScenarioContext, Measure}
-  alias Benchee.Utility.{RepeatN, Parallel}
+  alias Benchee.Benchmark.{Scenario, ScenarioContext, Measure, Hooks, RepeatedMeasurement}
+  alias Benchee.Utility.Parallel
   alias Benchee.Configuration
 
   @doc """
@@ -32,10 +32,10 @@ defmodule Benchee.Benchmark.Runner do
   # This will run the given scenario exactly once, including the before and
   # after hooks, to ensure the function can execute without raising an error.
   defp pre_check(scenario, scenario_context = %ScenarioContext{config: %{pre_check: true}}) do
-    scenario_input = run_before_scenario(scenario, scenario_context)
+    scenario_input = Hooks.run_before_scenario(scenario, scenario_context)
     scenario_context = %ScenarioContext{scenario_context | scenario_input: scenario_input}
-    _ = measure_iteration(scenario, scenario_context, Measure.Time)
-    _ = run_after_scenario(scenario, scenario_context)
+    _ = measure(scenario, scenario_context, Measure.Time)
+    _ = Hooks.run_after_scenario(scenario, scenario_context)
     nil
   end
 
@@ -68,32 +68,15 @@ defmodule Benchee.Benchmark.Runner do
   end
 
   defp measure_scenario(scenario, scenario_context) do
-    scenario_input = run_before_scenario(scenario, scenario_context)
+    scenario_input = Hooks.run_before_scenario(scenario, scenario_context)
     scenario_context = %ScenarioContext{scenario_context | scenario_input: scenario_input}
     _ = run_warmup(scenario, scenario_context)
     runtimes = run_runtime_benchmark(scenario, scenario_context)
     memory_usages = run_memory_benchmark(scenario, scenario_context)
-    run_after_scenario(scenario, scenario_context)
+    Hooks.run_after_scenario(scenario, scenario_context)
 
     {runtimes, memory_usages}
   end
-
-  defp run_before_scenario(
-         %Scenario{
-           before_scenario: local_before_scenario,
-           input: input
-         },
-         %ScenarioContext{
-           config: %{before_scenario: global_before_scenario}
-         }
-       ) do
-    input
-    |> run_before_function(global_before_scenario)
-    |> run_before_function(local_before_scenario)
-  end
-
-  defp run_before_function(input, nil), do: input
-  defp run_before_function(input, function), do: function.(input)
 
   defp run_warmup(
          scenario,
@@ -139,19 +122,6 @@ defmodule Benchee.Benchmark.Runner do
     do_benchmark(scenario, new_context, Measure.Memory, [])
   end
 
-  defp run_after_scenario(
-         %{
-           after_scenario: local_after_scenario
-         },
-         %{
-           config: %{after_scenario: global_after_scenario},
-           scenario_input: input
-         }
-       ) do
-    if local_after_scenario, do: local_after_scenario.(input)
-    if global_after_scenario, do: global_after_scenario.(input)
-  end
-
   defp measure_runtimes(scenario, context, run_time, fast_warning)
   defp measure_runtimes(_, _, 0.0, _), do: []
 
@@ -160,7 +130,7 @@ defmodule Benchee.Benchmark.Runner do
     :erlang.garbage_collect()
 
     {num_iterations, initial_run_time} =
-      determine_n_times(scenario, scenario_context, fast_warning)
+      RepeatedMeasurement.determine_n_times(scenario, scenario_context, fast_warning)
 
     new_context = %ScenarioContext{
       scenario_context
@@ -173,35 +143,6 @@ defmodule Benchee.Benchmark.Runner do
   end
 
   defp current_time, do: :erlang.system_time(:nano_seconds)
-
-  # If a function executes way too fast measurements are too unreliable and
-  # with too high variance. Therefore determine an n how often it should be
-  # executed in the measurement cycle.
-  @minimum_execution_time 10
-  @times_multiplier 10
-  defp determine_n_times(
-         scenario,
-         scenario_context = %ScenarioContext{
-           num_iterations: num_iterations,
-           printer: printer
-         },
-         fast_warning
-       ) do
-    run_time = measure_iteration(scenario, scenario_context, Measure.NativeTime)
-
-    if run_time >= @minimum_execution_time do
-      {num_iterations, adjust_for_iterations(run_time, num_iterations)}
-    else
-      if fast_warning, do: printer.fast_warning()
-
-      new_context = %ScenarioContext{
-        scenario_context
-        | num_iterations: num_iterations * @times_multiplier
-      }
-
-      determine_n_times(scenario, new_context, false)
-    end
-  end
 
   # `run_times` is kept separately from the `Scenario` so that for the
   # `parallel` execution case we can easily concatenate and flatten the results
@@ -223,7 +164,7 @@ defmodule Benchee.Benchmark.Runner do
   end
 
   defp do_benchmark(scenario, scenario_context, measurer, measurements) do
-    measurement = iteration_measurements(scenario, scenario_context, measurer)
+    measurement = measure(scenario, scenario_context, measurer)
     updated_context = %ScenarioContext{scenario_context | current_time: current_time()}
 
     do_benchmark(
@@ -238,131 +179,39 @@ defmodule Benchee.Benchmark.Runner do
   defp updated_measurements(nil, measurements), do: measurements
   defp updated_measurements(measurement, measurements), do: [measurement | measurements]
 
-  defp iteration_measurements(
-         scenario,
-         scenario_context = %ScenarioContext{
-           num_iterations: num_iterations
-         },
-         measurer
-       ) do
-    measurement = measure_iteration(scenario, scenario_context, measurer)
+  @doc """
+  Takes one measure with the given measurer.
 
-    adjust_for_iterations(measurement, num_iterations)
-  end
-
-  defp adjust_for_iterations(measurement, 1), do: measurement
-  defp adjust_for_iterations(measurement, num_iterations), do: measurement / num_iterations
-
-  defp measure_iteration(
-         scenario = %Scenario{function: function},
-         scenario_context = %ScenarioContext{
-           num_iterations: 1,
-         },
-         measurer
-       ) do
-    new_input = run_before_each(scenario, scenario_context)
+  Correctly dispatches based on the number of iterations to perform.
+  """
+  def measure(
+        scenario = %Scenario{function: function},
+        scenario_context = %ScenarioContext{
+          num_iterations: 1,
+        },
+        measurer
+      ) do
+    new_input = Hooks.run_before_each(scenario, scenario_context)
     function = main_function(function, new_input)
 
     {measurement, return_value} = measurer.measure(function)
 
-    run_after_each(return_value, scenario, scenario_context)
+    Hooks.run_after_each(return_value, scenario, scenario_context)
     measurement
   end
 
-  defp measure_iteration(
-         scenario,
-         scenario_context = %ScenarioContext{
-           num_iterations: iterations
-         },
-         measurer
-       )
-       when iterations > 1 do
-    # When we have more than one iteration, then the repetition and calling
-    # of hooks is already included in the function, for reference/reasoning see
-    # `build_benchmarking_function/2`
-    function = build_benchmarking_function(scenario, scenario_context)
-
-    {measurement, _return_value} = measurer.measure(function)
-
-    measurement
+  def measure(
+        scenario,
+        scenario_context = %ScenarioContext{
+          num_iterations: iterations
+        },
+        measurer
+      )
+      when iterations > 1 do
+    RepeatedMeasurement.measure(scenario, scenario_context, measurer)
   end
 
   @no_input Benchmark.no_input()
   defp main_function(function, @no_input), do: function
   defp main_function(function, input), do: fn -> function.(input) end
-
-  # Builds the appropriate function to benchmark. Takes into account the
-  # combinations of the following cases:
-  #
-  # * an input is specified - creates a 0-argument function calling the original
-  #   function with that input
-  # * number of iterations - when there's more than one iteration we repeat the
-  #   benchmarking function during execution and measure the the total run time.
-  #   We only run multiple iterations if a function is so fast that we can't
-  #   accurately measure it in one go. Hence, we can't split up the function
-  #   execution and hooks anymore and sadly we also measure the time of the
-  #   hooks.
-  defp build_benchmarking_function(
-         %Scenario{
-           function: function,
-           before_each: nil,
-           after_each: nil
-         },
-         %ScenarioContext{
-           num_iterations: iterations,
-           scenario_input: input,
-           config: %{after_each: nil, before_each: nil}
-         }
-       )
-       when iterations > 1 do
-    main = main_function(function, input)
-    # with no before/after each we can safely omit them and don't get the hit
-    # on run time measurements (See PR discussions for this for more info #127)
-    fn -> RepeatN.repeat_n(main, iterations) end
-  end
-
-  defp build_benchmarking_function(
-         scenario = %Scenario{function: function},
-         scenario_context = %ScenarioContext{num_iterations: iterations}
-       )
-       when iterations > 1 do
-    fn ->
-      RepeatN.repeat_n(
-        fn ->
-          new_input = run_before_each(scenario, scenario_context)
-          main = main_function(function, new_input)
-          return_value = main.()
-          run_after_each(return_value, scenario, scenario_context)
-        end,
-        iterations
-      )
-    end
-  end
-
-  defp run_before_each(
-         %{
-           before_each: local_before_each
-         },
-         %{
-           config: %{before_each: global_before_each},
-           scenario_input: input
-         }
-       ) do
-    input
-    |> run_before_function(global_before_each)
-    |> run_before_function(local_before_each)
-  end
-
-  defp run_after_each(
-         return_value,
-         %{
-           after_each: local_after_each
-         },
-         %{
-           config: %{after_each: global_after_each}
-         }
-       ) do
-    if local_after_each, do: local_after_each.(return_value)
-    if global_after_each, do: global_after_each.(return_value)
-  end
 end
